agent-registry/pyproject.toml
[project]
name = "agent-registry"
version = "0.1.0"
description = "A2A - MCP Resource Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "a2a-sdk>=0.3.5",
    "click>=8.1.8",
    "fastmcp>=1.0",
    "httpx>=0.28.1",
    "langchain>=0.3.26",
    "langchain-mcp-adapters>=0.0.9",
    "langgraph>=0.4.9",
    "mcp[cli]>=1.5.0",
    "nest-asyncio>=1.6.0",
    "networkx>=3.4.2",
    "numpy>=2.2.5",
    "pandas>=2.2.3",
    "pydantic>=2.11.4",
    "dashscope",
    "langchain_community>=0.3.26",
    "redis",
    "fastapi",
]

[project.scripts]
agent-registry = "agent_registry.server:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["agent_registry"]

[tool.pyright]
include = ["agent_registry"]
venvPath = "."
venv = ".venv"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = []

[tool.ruff]
line-length = 88
target-version = "py313"

[tool.uv]
dev-dependencies = []
index-url = "https://mirrors.aliyun.com/pypi/simple/"